// pages/helpers/DataTypes.d.ts
export interface DataFieldParams {
  data_fields: string;
}
export interface StartEndDate {
  startdate: number;
  enddate: number;
}

export interface StartEndYMD {
  startdateymd: string;
  enddateymd: string;
}

export interface QueryData {
  action: string;
  startDate: string;
  endDate: string;
}